policy: split out selectorcache.go
This is a noop change that just moves some functions and definitions
around. No code has been changed.
